Prepare the message parser for sharing and refactor subscribe response parsing
